First working example.
Correctly removes cars from the queue when they're clear of the 20-space highway area and the 4-space intersection--in other words, when the car's head is in Space 26 and its tail is in Space 25.

Generates cars based on a random probability of 0.5 (previous iterations used a 100% probability for the purposes of simplicity and testing the other, harder-to-implement features).

Outputs the number of cars in lane, and their locations, to the user at the end of the sim.

The green-light time and number of iterations is hardwired into the code for purposes of easier testing.  Later version will get green-light time, number of iterations and generation probability from the user.